feat: Keep up with the head of the chain. (#591)
# Description

- What does this PR do?
Our node needs to keep up with the head of the chain. In other words,
all announced blocks must be executed, and their states must be kept in
memory before being finalized. This PR implements the functionality to
do so. It also exposes a runtime API interface. This makes it more
readable and easily extensible. A few other changes were made along the
way that improved performance (async calls) and structured the project
in a better way.

- Why are these changes needed?
We need the latest state of the chain to validate transactions.

Fixes #574 and #573 

## Checklist:
- [X] I have read the [contributing
guidelines](https://github.com/LimeChain/Fruzhin/blob/dev/CONTRIBUTING.md).
- [X] My PR title matches the [Conventional Commits
spec](https://www.conventionalcommits.org/).
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [X] All new and existing tests passed.