refactor: Add NodeService and ServiceState interfaces (#688)
# Description
This PR aims to improve the overall structure of Fruzhin in 2 ways:
- Add a shared interface for services (sync and network, babe and
grandpa are to follow) `NodeService`. This enables services to share a
class signature and be started/stopped in a similar fashion. It also
helps separate what services are required and started for each type of
node.
- Add a shared interface for states (block, sync, epoch, grandpa and
transaction) `ServiceState`. Reasoning is similar to the service one,
but it also aims to centralize the initialization of the states as they
were not concise. It also prepares Fruzhin to follow the Runtime to
Consensus engine communication. (initializing states at genesis and
using consensus message to update)