fix: Implement a fix for polkaj's Hash256 and Hash512 serialization issues. (#721)
# Description
The Hash512 and Hash256 classes don't work with java's default
serialization logic. Considering they are nothing but wrapper around a
byte array field it is adequate to only serialize/deserialize the byte
array value of each instance. This PR implements changes needed in
Fruzhin after the update in polkaj

Fixes #717